Theme updates
Fix colors for tables that show status to set the foreground color in addition to the background color for accessibility.
Make display of availability accessible on all backgrounds
Update theming for search tools, and breadcrumbs so they can be customized.
Update theming for result labels and result values so they can be customized.
Update pagination controls to be bootstrap buttons for better accessibility and theming.