If pickupBranch is null, look at the users pickup location (LiDA)
In LiDA, if library only has 1 location, the user isn't prompted to select a pickup location, so it sends a null value for the parameter and an error is reported. I will also fix in LiDA code, but this gets a fix in the wild a bit faster and acts as a safety net.