EBSCOhost vs EDS
If both modules are enabled and settings/search settings exist for both EDS will be preferred and will be the only option that shows in the search source, explore more, and combined results. If EDS settings are set to "none" EBSCOhost will be the only option in these areas. This should make it easy for libraries to use either/or within a consortium that uses both