Evergreen Driver: fix crash when displaying certain fines
When a set of billings is associated with a checkout, the response
from open-ils.actor.user.transactions.have_charge.fleshed for the
underlying transaction will have additional keys with more details
about the loan and the title associated with the loan. The
billing summary record is the value of the 'transaction' key
in all cases, however.

To put it another way, the response from that method is an
array of hashes, each of which is guaranteed to have a 'transaction'
key, while most but not all will have additional keys.

This patch fixes the handling of the response and avoids a crash
that causes no fines to be displayed if one of them is associated
with a checkout.

Further enhancement is possible to include title, author, due date,
etc. in the Aspen fines display.