Account Linking Changes
When hitting 'enter' in the Add Account Link modal - the modal would close and the page would refresh and if there were an error, the message wouldn't show. Added an event listener for the user hitting enter that will just click the submit button so messaging shows as intended.
Changed "No" to "Manage Linked Accounts" in user message asking if user wants to continue allowing a new link