Materials Request Usage Dashboard
Changed column name in materials_request_usage from "locationId" to "libraryId" since that's what is being stored in that column
Updated other files to reflect this change in nomenclature
Dropped ability to choose location on this dashboard since it wasn't working correctly - we will now pull the stats for the user's home library. If they are an admin with no home library, stats will show for the active library.
Updated usage dashboard csv exports to reflect this change