Evergreen indexer: ignore items that are not OPAC visible
This patch ensures that if an item is marked as not OPAC
visible in any of the four ways that this can be done in
Evergreen, that it is excluded when refreshing a record
via supercat. This needs to be matched with excluding OPAC-hidden
items from the periodic MARC exports from Evergreen.

Signed-off-by: Galen Charlton <gmc@equinoxOLI.org>
Signed-off-by: Jason Boyer <jboyer@equinoxOLI.org>