Evergreen: enable password recovery
This follows the same flow as Koha.

The main Evergreen setup task is adjusting the password reset email
Action Trigger event definition to direct the user to

https://ASPEN/MyAccount/PasswordRecovery/?uniqueKey=[% target.uuid %]

In cases where only a subset of an Evergreen consortium is using Aspen,
the event defintion may need to check the requesting patrons' home
library to decide whether to direct the password recovery completion to
Aspen or to the Evergreen OPAC.

Since Evergreen enforces its password complexity requirements as part
of the recovery process, Evergreen's password strength rules and
Aspen's should be configured to be in alignment.

Signed-off-by: Galen Charlton <gmc@equinoxOLI.org>
Signed-off-by: Jason Boyer <jboyer@equinoxOLI.org>