Browse Category Group Edit
Updated the Browse Category Groups page to check for the user's home library as well as their permissions so if they have "Administer Library Browse Categories" they will only be able to edit browse categories shared with everyone or with their library. They also cannot select a new browse category if it is not shared with everyone or their library, they can only delete it from the group.
Updated release notes