feat: Create cookie policy html database update and system variable.
This will allow you to specify the HTML that will go into the cookie policy modal that appears when an end user requests more info.
This includes passing the HTML from the syspref to smarty to be used in
the template file.

Additionally move relevant system variables to new Data protection
section.
Variables included:
 - Cookie Storage consent
 - Cookie policy HTML
 - Track IP Addresses