ensure that Square is passed a string for idempotency_key
Depending on PHP and MySQL driver versions, PHP is increasingly
more likely to return numeric columns as PHP integers or floats.

For example, PHP 8.1+ will now do this for emulated prepared
statements.

Since Square requires that the idempotency_key be a string,
but Aspen uses the numeric user_payments ID for the purpose,
this patch ensures that it gets cast to a string.

Concretely, this patch fixes a bug where payments can fail to be
created in Square with the following error (as found in Square's
API logs):

{
  "errors": [
    {
      "code": "EXPECTED_STRING",
      "detail": "Expected a string value (line 1, character 20)",
      "field": "idempotency_key",
      "category": "INVALID_REQUEST_ERROR"
    }
  ]
}

Signed-off-by: Galen Charlton <gmc@equinoxOLI.org>