Get Barcode and Call Numbers of checkouts in Evergreen
Add a getCallNumberForCopy() method to the Evergreen driver to
retrieve the call number label for a mapped copy.

Refactor the Evergreen driver's getCheckouts() method to use the
open-ils.circ.actor.user.checked_out API, to build the checkout object
itself from the returned values, and to use the getCallNumberForCopy()
method to get the copy's call number for display.

The loadCheckoutData() method remains untouched because it is
currently used elsewhere.

Signed-off-by: Jason Stephenson <jason@sigio.com>