fix: summon results filtering fix
Adjust code so that if not on page one of results when a filter is applied,
the new results set begins on page one.
Fix bug: if current result numerical value is greater than the number of
results available in a facet filter, the facet filter did not apply
when clicked.