fix: remove trailing slash
The output path for the reindexer artifact contained a trailing slash.
Upon the building process, this resulted in a malformed path containing
a double slash, which meant that the older reindexer.jar file was not
getting rebuilt when using the IntelliJ rebuild artifact feature.
Removing the trailing slash resolves the issue.