fix: guard clause handles null values
During the process of updating user preferences,
$user->getCatalogDriver() may return null, thus the need for a guard
clause in $user->updateHomeLibrary. Since this guard clause makes it
possible for $user->updatedHomeLibrary to return null as well, an
additional guard clause is added to the launch function for
'MyPreferences'.

This sould prevent the 'Member function updateHomeLibrary called on
Null' from being displayed on the OPAC upon attempts to save user
preferences.