fix: extract branch name
The branch name improperly extracted, leading to
$branchList[$branchName]->showInSearchFacet systematically evaluated to
false for scopes where the length of the scope code was not 4. All
locations for that scope would be filtered out, resulting in empty
facets which would therefore not be rendered.

This fix replaces the hardcoded offset in the substr() method by the
length of the scope + 1 to account for the #, allowing for proper branch
name extraction, and appropriate filtering.