feat: build CSV file
This sets up the CSV file and triggers its download. It will contain the
column headers of the raw data table. The part setting up the output to
the browser is directly based on the build CSV function in
sys/UserLists/UserList.php, and only the filename has been changed. The
CSV is populated with as many rows of data as found within the raw usage
data table.