Remove old pre-production Koha volumes code
Any libraries using the pre-release version of Koha's item groups ( i.e. volumes ) has since been upgraded past 22.11.
This code should be safe to remove now as it relates to a version of a feature that was never part of community Koha.

Test Plan:
No changes in behavior should be noted

Closes Aspen-Discovery/aspen-discovery#1990