Implement dependent issues for GitHub Issues
We are currently using GitHub Issues for issue tracking While we
are deciding on a long term issue tracker for the community ( Issues or Jira ),
In the meantime it would be nice if we could use the dependent-issues
action to deal with issue dependencies in the community.

Once this branch is merged, setting a dependency requires only that
one leave a comment on the issue in the format:

Depends on Aspen-Discovery/aspen-discovery#XXXX

Where XXXX is the dependent issue number.
Any number of "Depends on" lines can exist in one or more comments.

https://github.com/marketplace/actions/dependent-issues#dependent-issues

Closes Aspen-Discovery/aspen-discovery#1993