Only attempt to log in and push to container repos if the commit is pushed to the official repo
When a developer pushes a branch of Aspen Discovery to their own repo to
create a pull request, it triggers the building of Docker images.
The building of the images itself is good to find issues in the build
process, but the actions fail because the the actions can only log
in on the official repo.

This change resolves this be only logging in to the registries if the
commit is pushed to the official repo, while still allowing the docker
images to be build for testing purposes.