Get volumes from MARC record when volumes.txt does not exist
Sirsi no longer provides volume information in a text file for new Aspen libraries but does put the volume number in the volume subfield of the item in the MARC record. They do not provide the call number key, which is required to place a hold, so save the volume information in the ils_volume_info table but prefix the volume_id field with LOOKUP so that the call number key can be looked up later via API when placing the hold.  Existing Aspen libraries continue to use the volumes.txt file.  ByWater Ticket #134021