Issue 2026: Add GH Action to check for changes in site creation that may need changes in docker
This action attempts to do basic checking of changes to createSite.php
that would need changes to Docker files, and then checks to see if
any files in the docker directory have been changes.

First, the action checks for any changes to createSiteTemplate.ini.
If these changes are detected it will then check to see if anything
in the docker directory has been changed. If nothing has been changed,
the step will fail. Otherwise the step will pass.

Test Plan:
1) Pick this commit to a branch
2) Modify install/createSiteTemplate.ini
3) Commit this change as a new commit
4) Push this branch to GitHub
5) Note the action fails!
6) Modify some file in the docker dir
7) Commit this changes as well
8) Push the updated branch to GitHub
9) That action should now pass!

Closes Aspen-Discovery/aspen-discovery#2026