Fix the check_tabs_vs_spaces GitHub Action
It appears that quoting a regex in bash causes it to use it as a string
literal so the grep chained to the git diff command never worked.
In addition the regular expression would have only tested the first line
of the diff instead of each line if the diff.

This commit updates that step to loop over the results of the diff and
test each one with the corrected regex. It will now display a list of
affected file at the end of the run if any are detected.

Test Plan:
1) Modify a file to add spaces instead of tabs at the start of a line
2) Push to your repo
3) Note the step fails!
4) Replace the spaces with tabs
5) Push to your repo
6) Note the step passes!