fix: remove 'hide from search facet'
This removes the 'Show This Branch in Available At and Owning Locations
Facets' setting and filter in order to resolve the bug created by the
introduction of this feature.

Bug: branches disappearing from location-related search facets

Cause: the filter ultimately (and incorrectly) compares $displayName to
$facetLabel (as extracted from $facetValue) when attempting to access a
location's 'showInSearchFacet' property and check its value.
If those don't match, no branch is found, the check fails, and the
branch is not added to the facet

Test plan:

to replicate the bug:
- run a search and take notes of the branches appearing in the Available
At search facet
- for one of these branches, set  the  'Library System Facet Label' to a
value that does not match that Library System's display name
- run a search again
- notice the branch has disappeared from the search facet

apply the patch, then go through the test plan again: the branch should
be display in the serach facet correctly, no matter what its label is
set to.