Bug fix for error submitting contact info form
Some Symphony libraries use a different field than PHONE for phone numbers (DAYPHONE or HOMEPHONE) and this caused an error on the Contact Information page (Expected a resource of type '/policy/patronAddress1' for the field 'address1[8]').  This change adds a setting for default phone field (which defaults to PHONE).  The Contact Information page will use the default phone field name for phone updates.  When updating patron data in the Aspen database, if either DAYPHONE or HOMEPHONE exists but PHONE does not, they will be saved in the Aspen phone field.