Add display name to User table when adding linked accounts via the API
If an account that belonged to someone who had never logged into Aspen before was added as a Linked Account via the API (ie, in LiDA), the user was being created in the database, but displayName wasn't getting added.  This made it so that the linked account's name was missing on pages like the Library Card page.