fix: Aspen Error messages are shown
When an Aspen error is thrown for EBSCO EDS search
and the searchError.tpl file is displayed, instead
of the original error message, this php error is
shown:

Cannot use object of type AspenError as array

Test plan:
- navigate to Aspen Administration > EBSCO EDS >
Settings
- create a new Setting, make sure to use invalid
credentials and save
- navigate to Aspen Administration > Primary
Configuration > Library Systems and select your
current Library
- set its EDS Setting to the setting you just
created
- run a search in Articles and Databases

Before this commit:
The error message is uninformative (Cannot use
object of type AspenError as array).

After this commit:
The error message refers to a specific cause.