feat: error messages are specific
Error messages should narrow down the cause
to whether the pertain to authentication, issues
loading data from EDS, or processing the search,
and do so without risking to expose any data that
should not be made publicly visible.

Test plan:

- navigate to Aspen Administration > EBSCO EDS >
Settings
- create a new Setting, make sure to use invalid
credentials and save
- navigate to Aspen Administration > Primary
Configuration > Library Systems and select your
current Library
- set its EDS Setting to the setting you just
created
- log out
- run a search in Articles and Databases

Before this commit, the error message will
indicate that the EDS search failed. After this,
it will indicate that it failed due to an
EDS authentication issue.

Note: this also removes the possibility for errors
thrown during the search processing to be show to
the end user directly - they are logged instead.