DIS-76 feat: handle payment attempt outcome
Once the user has completed their payment on the
HeyCentric platform, they are redirected via the
RURL supplied to HeyCentric. The record of the
relevant user payment(s) are updated in Aspen db
to reflect the outcome of the payment attempt.

Additionally, if the payment attempt was
successful, the debt in the ILS Aspen is set to
use will also be updated.

Note: this was built and tested with Koha as the
ILS.